% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural-earth.R
\docType{data}
\name{borders}
\alias{borders}
\title{Country borders}
\format{
\describe{
\item{country}{ISO2 country code}
\item{long,lat}{Location}
\item{group}{Grouping variable need to separate islands etc.}
}
}
\source{
http://www.naturalearthdata.com/downloads/50m-cultural-vectors/
}
\usage{
borders
}
\description{
Admin 0 (country) borders for the region
}
\keyword{datasets}
