% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasa.R
\docType{data}
\name{elev}
\alias{elev}
\title{Elevation.}
\format{
\describe{
\item{lat,long}{Location of measurement. Evenly spaced 24 by 24 spatial
  grid from longitude 113.8W to 56.2W and from latitude 36.2N to 21.2S}
\item{elev}{Height of location above reference plane (in m)}
}
}
\source{
\url{http://stat-computing.org/dataexpo/2006/}
}
\usage{
elev
}
\description{
As of Jan 1998.
}
\keyword{datasets}
